Update code_controller.dart
The error that the patternMap variable is not updated is not resolved with the copyWith method. The source of the error is line 86: patternList.addAll(patternMap!.keys.map((e) => r'($e)')); the patternMap variable could not be updated because r'($e)' was written as '($e)' in this line. The error has been fixed.